backend/requirements.txt
# Core framework
fastapi==0.115.2
uvicorn[standard]==0.30.6

# Async database (PostgreSQL)
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0

# Redis async client
redis==5.0.8

# HTTP client (for Hugging Face API)
httpx==0.27.2

# Metrics & monitoring
prometheus-client==0.20.0

# Security & Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Data validation
pydantic==2.9.2
pydantic-settings==2.5.2

# Optional: helpful tools
python-multipart==0.0.9   # for future file uploads (if you ever add avatars)
